+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - do some imports closer to where they are used #1810 @williballenthin
 - binja: fix and simplify stack string detection code after binja 4.0 @xusheng6
 - binja: add support for forwarded export #1646 @xusheng6
+- cape: support more report formats #2035 @mr-tz
 
 
 ### capa explorer IDA Pro plugin